% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easytimingr.R
\name{absTOsingleTimes}
\alias{absTOsingleTimes}
\alias{absTOsingleTimes.RuntimeC}
\title{Finalize the RuntimeC-Object}
\usage{
absTOsingleTimes(object, forceFinal)

\method{absTOsingleTimes}{RuntimeC}(object, forceFinal = FALSE)
}
\arguments{
\item{object}{RuntimeC-object}

\item{forceFinal}{Forces adding the sumary and calculating the relative times}
}
\value{
object with 2 summary rows and all rows in relative times.
}
\description{
Adds a summary of all times in RuntimeC and calculates the relative times for
each function (Instead of the absolute time of proc.time())
Should be done only once, multiple times yield in pointless matrix. Therefore,
usually it can be done only once, except you force it by setting forceFinal=TRUE.
}
\examples{
library(basicscriptsGG)
rt_object <- RuntimeC()
rt_object <- add.timepoint(rt_object, "startpoint")			# add new timepoint NOW
rt_object <- add.timepoint(rt_object, "fun1")			# add new timepoint NOW
rt_object <- absTOsingleTimes(rt_object)	# subtract the first for (starting point)
}
