% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easytimingr.R
\name{add.timepoint}
\alias{add.timepoint}
\alias{add.timepoint.RuntimeC}
\title{Adding timepoints}
\usage{
add.timepoint(object, name, scripttime)

\method{add.timepoint}{RuntimeC}(object, name, scripttime)
}
\arguments{
\item{object}{RuntimeC-class object}

\item{name}{Name of time point}

\item{scripttime}{Time for one script, given by "proc.time() - proc.time()"}
}
\description{
Adds a named timepoint to object
}
\examples{
library(basicscriptsGG)
rt_object <- RuntimeC()
rt_object <- add.timepoint(rt_object, "useless", proc.time()-proc.time())	# add new timepoint (with zero time)
t1 <- proc.time()
Sys.sleep(1)		# someFunction
t2 <- proc.time()
rt_object <- add.timepoint(rt_object, "someFunction", t2-t1)	# add new timepoint (with zero time)
rt_object <- add.timepoint(rt_object, "someFunction", t2-t1)	# add new timepoint (with zero time)
rt_object <- add.timepoint(rt_object, "someFunction", t2-t1)	# add new timepoint (with zero time)

}
