% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catt.R
\name{define.catt}
\alias{define.catt}
\title{A modified cat()}
\usage{
define.catt(ntab = NULL, nspace = NULL)
}
\arguments{
\item{ntab}{How many tabs indent}

\item{nspace}{How many spaces indent}
}
\description{
A modified cat() (therefore catt()), to have empty space before output generally
If defince.catt() is done as wished, catt's usage is as cat().
}
\examples{
var <- "testtext\\n"
catt <- define.catt(nspace=2) # from now on, all catt("anyText") are indented by two spaces
catt(var, "\\n", "  ", var)
catt <- define.catt() # reset to 0 indent

# defines overwrite each other
catt(var)
catt <- define.catt(nspace=2)
catt(var)
catt <- define.catt(nspace=3)
catt(var)

# combinations are not possible in a usefull way.
catt <- define.catt(nspace=2, ntab=3)
catt(var)
catt <- define.catt()
}
