% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst.load.packages.R
\name{inst.load.packages}
\alias{inst.load.packages}
\title{Loads a package called "package" and installs it if necessary}
\usage{
inst.load.packages(package, silent = FALSE, inst.bioclite = FALSE)
}
\arguments{
\item{package}{The package you want to load}

\item{silent}{If "TRUE", no output is generated}

\item{instBiocLite}{If it is a package from bioconductor, set it to TRUE}
}
\value{
loads the "package"
}
\description{
Loads a package called "package" and installs it if necessary
}
\examples{
inst.load.packages("ggplot2") # loads (& installs) ggplot2
inst.load.packages("ggplot2", silent=TRUE)

# if package not already installed it also installs it with inst.bioclite=TRUE
inst.load.packages("vioplot", inst.bioclite=TRUE)

# if bioconductor package should be installed, inst.bioclite MUST be =TRUE
inst.load.packages("NanoStringQCPro", inst.bioclite=TRUE)
}
