% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.genecards.R
\name{links.genecards}
\alias{links.genecards}
\title{Gene list to links from genecards.org}
\usage{
links.genecards(genelist, urlfile = "GeneCards.genelist.txt")
}
\arguments{
\item{genelist}{vector of genes}

\item{urlfile}{File where the links should be stored}
}
\value{
urlfile (writes it into workingDIR)
}
\description{
Gene list to links from genecards.org
}
\examples{
genes <- c("MYC", "NLRP11")
links.genecards(genes)
links.genecards(genes, urlfile="links.genecards.txt")
}
