% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.textable.R
\name{make.textable}
\alias{make.textable}
\title{Booktabs table from matrix
Make a latex table to use with the booktaps package from any array}
\usage{
make.textable(array0, value.formatstring = NA, title = NA, label = NA,
  cutfunction = function(x) {     return(FALSE) }, rowcol = "red!10",
  sisetup = FALSE)
}
\arguments{
\item{array0}{An array}

\item{value.formatstring}{How should a single value of the array be formatted? (Not the row/colnames)}
}
\description{
Booktabs table from matrix
Make a latex table to use with the booktaps package from any array
}
\examples{
arr <- matrix(rnorm(20), ncol=4, nrow=5)
rownames(arr) <- paste0("Row", 1:5)
colnames(arr) <- paste0("Col", 1:4)
make.textable(arr)
make.textable(arr, title = "abc")
make.textable(arr, title = "abc", label = "testlabel")
make.textable(arr, value.formatstring = "\%5.3e")
make.textable(arr, value.formatstring = "$\%5.3e$")
make.textable(arr, cutfunction=function(x){x>1})
make.textable(arr, value.formatstring = "$\%5.3e$")

# to save into a file:
sink("testfile.txt")
make.textable(arr, value.formatstring = "\%5.3f")
sink()
}
