% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easytimingr.R
\name{RuntimeC}
\alias{RuntimeC}
\title{Constructor for S3 RuntimeC-Class}
\usage{
RuntimeC()
}
\value{
RuntimeC-object
}
\description{
Invokes a RuntimeC-Object holding a matrix with 5 cols and 0 rows where future
time-points are stored
}
\examples{
rt_object <- RuntimeC()
}
