% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.percentage.R
\name{sample.percentage}
\alias{sample.percentage}
\title{Divides vector0 into random "length(percentages)" parts}
\usage{
sample.percentage(vector0, percentages = c(0.5, 0.2, 0.3), seed0 = NA)
}
\arguments{
\item{vector0}{Vector of numbers, strings, etc. ("sample" must handle it)}

\item{percentages}{Determines how many groups and how many (percentagewise)
samples should be per group}

\item{seed0}{A seed for reproducability, seed will NOT be the same
for each group (but reproducible: seed <- seed0 + #group)}
}
\value{
List of sorted, by "percentages" randomly divided vector0
}
\description{
Divides vector0 into random "length(percentages)" parts
}
\examples{

words  <- c("a","b","c","d")
sample.percentage(words)


vec <- 0:99
sample.percentage(vec)

sample.percentage(vec, percentages=c(0.5,0.5))

sample.percentage(vec, percentages=c(0.1, 0.1, 0.1, 0.7))

}
