% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.addtime.R
\name{source.addtime}
\alias{source.addtime}
\title{A personal adaptation of source()}
\usage{
source.addtime(scriptpath, runtimecobject)
}
\arguments{
\item{scriptpath}{path to a script to execute}

\item{runtimecobject}{The RuntimeC-object where the timepoint should be added.}
}
\description{
This source.addtime() makes use of the RuntimeC-Class and includes making
a time-point aswell as sourcing the script.
Advantage: one-liner and the called script has access to its name.
The script will be sourced automatically in .GobalEnv (according to stackoverflow)
I also tested this and it seems to work.
}
\examples{
library(basicscriptsGG)
A_RUNTIME <- RuntimeC()		# initialize runtime-array
# source.addtime("ANY.Scriptpath", A_RUNTIME)
}
