% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easytimingr.R
\name{summarizeRT}
\alias{summarizeRT}
\alias{summarizeRT.RuntimeC}
\title{Finalize the RuntimeC-Object}
\usage{
summarizeRT(object, forceFinal)

\method{summarizeRT}{RuntimeC}(object, forceFinal = FALSE)
}
\arguments{
\item{object}{RuntimeC-object}

\item{forceFinal}{Forces adding the summary and calculating the relative times}
}
\value{
object with 2 summary rows and all rows hold the scripttime of the rownames()
}
\description{
Adds a summary of all times in RuntimeC
}
\examples{
library(basicscriptsGG)
rt_object <- RuntimeC()
rt_object <- add.timepoint(rt_object, "useless", proc.time()-proc.time())	# add new timepoint (with zero time)
t1 <- proc.time()
Sys.sleep(1)		# someFunction
t2 <- proc.time()
rt_object <- add.timepoint(rt_object, "someFunction", t2-t1)	# add new timepoint (with zero time)
rt_object <- summarizeRT(rt_object)	# subtract the first for (starting point)
}
