% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easytimingr.R
\name{write}
\alias{write}
\alias{write.default}
\alias{write.RuntimeC}
\title{Write function}
\usage{
write(x, ...)

\method{write}{default}(x, file = "data", ncolumns = if (is.character(x)) 1
  else 5, append = FALSE, sep = " ")

\method{write}{RuntimeC}(obj, file)
}
\arguments{
\item{object}{See ?base::write}
}
\description{
Just the normal write function made generic and RuntimeC-support added
}
\examples{
# See ?base::write. For RuntimeC:
library(basicscriptsGG)
rt_object <- RuntimeC()
rt_object <- add.timepoint(rt_object, "startpoint")			# add new timepoint NOW
rt_object <- add.timepoint(rt_object, "fun1")			# add new timepoint NOW
rt_object <- absTOsingleTimes(rt_object)	# subtract the first for (starting point)
write(rt_object, "test.txt")

}
