% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easytimingr.R
\name{add.timepoint}
\alias{add.timepoint}
\alias{add.timepoint.RuntimeC}
\title{Adding timepoints}
\usage{
add.timepoint(object, name)

\method{add.timepoint}{RuntimeC}(object, name)
}
\arguments{
\item{object}{RuntimeC-class object}

\item{name}{Name of time point}
}
\description{
Adds a named timepoint to object
}
\examples{
library(basicscriptsGG)
rt_object <- RuntimeC()
rt_object <- add.timepoint(rt_object, "startpoint")			# add new timepoint NOW
rt_object <- add.timepoint(rt_object, "fun1")			# add new timepoint NOW
Sys.sleep(1) # placeholder for ultra cool function/script
rt_object <- add.timepoint(rt_object, "fun2")			# add new timepoint NOW
Sys.sleep(0.1)
rt_object <- add.timepoint(rt_object, "fun3")			# add new timepoint NOW
Sys.sleep(.2)

}
